sched/fair: improve big task placement
When a task that should run on the highest cluster is being placed and
if all the cpus in the highest cluster are either not available or
cannot accommodate this task, we could end up placing this task on
the smallest cpu.

Instead place this task on the most spare capacity cpu.

Change-Id: I10256918a7d6f016e7fce31af61c314b8784e22d
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Yousef Algadri <yusufgadrie@gmail.com>
Signed-off-by: palaych <palaychm@yandex.ru>